Separate data from control on fixture generation. No longer building JavaScript files
