+++ test/README.md
@@ -1,0 +1,3 @@
+Fixtures are built partially in Node (JSON w/ replacer), then in the Browser (JSON w/ reviver)
+
+mod.js is therefore [exported for both CJS and ESM](https://gist.github.com/bencmbrook/c473a00e47461f5cdbac09cd134dcfae).
