+++ test/fixture.test.js
@@ -1,262 +1,5 @@
-const mod = () =>
-  ((types, b64) => ({
-    replacer(key) {
-      const val = this[key];
-      return ArrayBuffer.isView(val)
-        ? { $d: b64.encode(new Uint8Array(val)) }
-        : val;
-    },
-    reviver: (key, val) =>
-      val === null && val !== 'object'
-        ? val
-        : val.$d
-        ? new Blob(b64.decode(val))
-        : val,
-  }))(
-    [],
-    (() => {
-      const f = [
-        65,
-        66,
-        67,
-        68,
-        69,
-        70,
-        71,
-        72,
-        73,
-        74,
-        75,
-        76,
-        77,
-        78,
-        79,
-        80,
-        81,
-        82,
-        83,
-        84,
-        85,
-        86,
-        87,
-        88,
-        89,
-        90,
-        97,
-        98,
-        99,
-        100,
-        101,
-        102,
-        103,
-        104,
-        105,
-        106,
-        107,
-        108,
-        109,
-        110,
-        111,
-        112,
-        113,
-        114,
-        115,
-        116,
-        117,
-        118,
-        119,
-        120,
-        121,
-        122,
-        48,
-        49,
-        50,
-        51,
-        52,
-        53,
-        54,
-        55,
-        56,
-        57,
-        45,
-        95,
-        61,
-      ];
-      const h = [
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        null,
-        62,
-        null,
-        62,
-        null,
-        63,
-        52,
-        53,
-        54,
-        55,
-        56,
-        57,
-        58,
-        59,
-        60,
-        61,
-        null,
-        null,
-        null,
-        64,
-        null,
-        null,
-        null,
-        0,
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9,
-        10,
-        11,
-        12,
-        13,
-        14,
-        15,
-        16,
-        17,
-        18,
-        19,
-        20,
-        21,
-        22,
-        23,
-        24,
-        25,
-        null,
-        null,
-        null,
-        null,
-        63,
-        null,
-        26,
-        27,
-        28,
-        29,
-        30,
-        31,
-        32,
-        33,
-        34,
-        35,
-        36,
-        37,
-        38,
-        39,
-        40,
-        41,
-        42,
-        43,
-        44,
-        45,
-        46,
-        47,
-        48,
-        49,
-        50,
-        51,
-        null,
-        null,
-      ];
-      return {
-        decode(a) {
-          let b = a.length % 4;
-          b && (a += Array(5 - b).join('='));
-          b = -1;
-          const f = new ArrayBuffer((a.length / 4) * 3);
-          let d;
-          const e = new Uint8Array(f);
-          let c = 0;
-          for (d = a.length; ++b < d; ) {
-            let g = h[a.charCodeAt(b)];
-            let k = h[a.charCodeAt(++b)];
-            e[c++] = (g << 2) | (k >> 4);
-            g = h[a.charCodeAt(++b)];
-            if (g === 64) break;
-            e[c++] = ((k & 15) << 4) | (g >> 2);
-            k = h[a.charCodeAt(++b)];
-            if (k === 64) break;
-            e[c++] = ((g & 3) << 6) | k;
-          }
-          return new Uint8Array(f, 0, c);
-        },
-        encode(a) {
-          for (
-            var b = -1,
-              h = a.length,
-              d = new Uint8Array(new ArrayBuffer(Math.ceil((4 * h) / 3))),
-              e = 0;
-            ++b < h;
+import mod from './mod.js';
+import fixtures from './fixtures.json';
 
-          ) {
-            let c = a[b];
-            const g = a[++b];
-            d[e++] = f[c >> 2];
-            d[e++] = f[((c & 3) << 4) | (g >> 4)];
-            isNaN(g)
-              ? ((d[e++] = f[64]), (d[e++] = f[64]))
-              : ((c = a[++b]),
-                (d[e++] = f[((g & 15) << 2) | (c >> 6)]),
-                (d[e++] = f[isNaN(c) ? 64 : c & 63]));
-          }
-          return new TextDecoder().decode(d);
-        },
-      };
-    })(),
-  );
 const { reviver } = mod();
-export default JSON.parse(
-  '{"all-stream.zip":{"$d":"UEsDBAoACAAAAJxs8ToAAAAADAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsHCOPllbAMAAAADAAAAFBLAwQKAAAAAACxbPE6AAAAAAAAAAAAAAAABwAAAGltYWdlcy9QSwMECgAIAAAAFWfqOgAAAAApAAAAKQAAABAAAABpbWFnZXMvc21pbGUuZ2lmR0lGODdhBQAFAIACAAAAAP_eACwAAAAABQAFAAACCIwPkWerClIBADtQSwcIPD_riikAAAApAAAAUEsBAh4DCgAIAAAAnGzxOuPllbAMAAAADAAAAAkAAAAAAAAAAAAAAKSBAAAAAEhlbGxvLnR4dFBLAQIeAwoAAAAAALFs8ToAAAAAAAAAAAAAAAAHAAAAAAAAAAAAEADtQUMAAABpbWFnZXMvUEsBAh4DCgAIAAAAFWfqOjw_64opAAAAKQAAABAAAAAAAAAAAAAAAKSBaAAAAGltYWdlcy9zbWlsZS5naWZQSwUGAAAAAAMAAwCqAAAAzwAAAAAA"},"all.7zip.zip":{"$d":"UEsDBAoAAAAAAGJ-_zrj5ZWwDAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsDBAoAAAAAAGJ-_zoAAAAAAAAAAAAAAAAHAAAAaW1hZ2VzL1BLAwQKAAAAAABifv86PD_riikAAAApAAAAEAAAAGltYWdlcy9zbWlsZS5naWZHSUY4N2EFAAUAgAIAAAAA_94ALAAAAAAFAAUAAAIIjA-RZ6sKUgEAO1BLAQIUAAoAAAAAAGJ-_zrj5ZWwDAAAAAwAAAAJAAAAAAAAAAAAIAAAAAAAAABIZWxsby50eHRQSwECFAAKAAAAAABifv86AAAAAAAAAAAAAAAABwAAAAAAAAAAABAAAAAzAAAAaW1hZ2VzL1BLAQIUAAoAAAAAAGJ-_zo8P-uKKQAAACkAAAAQAAAAAAAAAAAAIAAAAFgAAABpbWFnZXMvc21pbGUuZ2lmUEsFBgAAAAADAAMAqgAAAK8AAAAAAA"},"all.windows.zip":{"$d":"UEsDBBQAAAAAAGJ-_zrj5ZWwDAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsDBBQAAAAIAGJ-_zo8P-uKJQAAACkAAAAQAAAAaW1hZ2VzL3NtaWxlLmdpZnP3dLMwT2RlYGVoYGIAgv_3GHRANEiEgYmjh39i-mquIEYGawBQSwECFAAUAAAAAABifv864-WVsAwAAAAMAAAACQAAAAAAAAABACAAAAAAAAAASGVsbG8udHh0UEsBAhQAFAAAAAgAYn7_Ojw_64olAAAAKQAAABAAAAAAAAAAAAAgAAAAMwAAAGltYWdlcy9zbWlsZS5naWZQSwUGAAAAAAIAAgB1AAAAhgAAAAAA"},"all.zip":{"$d":"UEsDBAoDAAAAAJxs8Trj5ZWwDAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsDBAoDAAAAALFs8ToAAAAAAAAAAAAAAAAHAAAAaW1hZ2VzL1BLAwQKAwAAAAAVZ-o6PD_riikAAAApAAAAEAAAAGltYWdlcy9zbWlsZS5naWZHSUY4N2EFAAUAgAIAAAAA_94ALAAAAAAFAAUAAAIIjA-RZ6sKUgEAO1BLAQIUAwoDAAAAAJxs8Trj5ZWwDAAAAAwAAAAJAAAAAAAAAAAAIICkgQAAAABIZWxsby50eHRQSwECFAMKAwAAAACxbPE6AAAAAAAAAAAAAAAABwAAAAAAAAAAABCA7UEzAAAAaW1hZ2VzL1BLAQIUAwoDAAAAABVn6jo8P-uKKQAAACkAAAAQAAAAAAAAAAAAIICkgVgAAABpbWFnZXMvc21pbGUuZ2lmUEsFBgAAAAADAAMAqgAAAK8AAAAAAA"},"all_appended_bytes.zip":{"$d":"UEsDBAoDAAAAAJxs8Trj5ZWwDAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsDBAoDAAAAALFs8ToAAAAAAAAAAAAAAAAHAAAAaW1hZ2VzL1BLAwQKAwAAAAAVZ-o6PD_riikAAAApAAAAEAAAAGltYWdlcy9zbWlsZS5naWZHSUY4N2EFAAUAgAIAAAAA_94ALAAAAAAFAAUAAAIIjA-RZ6sKUgEAO1BLAQIUAwoDAAAAAJxs8Trj5ZWwDAAAAAwAAAAJAAAAAAAAAAAAIICkgQAAAABIZWxsby50eHRQSwECFAMKAwAAAACxbPE6AAAAAAAAAAAAAAAABwAAAAAAAAAAABCA7UEzAAAAaW1hZ2VzL1BLAQIUAwoDAAAAABVn6jo8P-uKKQAAACkAAAAQAAAAAAAAAAAAIICkgVgAAABpbWFnZXMvc21pbGUuZ2lmUEsFBgAAAAADAAMAqgAAAK8AAAAAAEhlbGxvIFdvcmxkCg"},"all_missing_bytes.zip":{"$d":"bGxvLnR4dEhlbGxvIFdvcmxkClBLAwQKAwAAAACxbPE6AAAAAAAAAAAAAAAABwAAAGltYWdlcy9QSwMECgMAAAAAFWfqOjw_64opAAAAKQAAABAAAABpbWFnZXMvc21pbGUuZ2lmR0lGODdhBQAFAIACAAAAAP_eACwAAAAABQAFAAACCIwPkWerClIBADtQSwECFAMKAwAAAACcbPE64-WVsAwAAAAMAAAACQAAAAAAAAAAACCApIEAAAAASGVsbG8udHh0UEsBAhQDCgMAAAAAsWzxOgAAAAAAAAAAAAAAAAcAAAAAAAAAAAAQgO1BMwAAAGltYWdlcy9QSwECFAMKAwAAAAAVZ-o6PD_riikAAAApAAAAEAAAAAAAAAAAACCApIFYAAAAaW1hZ2VzL3NtaWxlLmdpZlBLBQYAAAAAAwADAKoAAACvAAAAAAA"},"all_prepended_bytes.zip":{"$d":"SGVsbG8gV29ybGQKUEsDBAoDAAAAAJxs8Trj5ZWwDAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsDBAoDAAAAALFs8ToAAAAAAAAAAAAAAAAHAAAAaW1hZ2VzL1BLAwQKAwAAAAAVZ-o6PD_riikAAAApAAAAEAAAAGltYWdlcy9zbWlsZS5naWZHSUY4N2EFAAUAgAIAAAAA_94ALAAAAAAFAAUAAAIIjA-RZ6sKUgEAO1BLAQIUAwoDAAAAAJxs8Trj5ZWwDAAAAAwAAAAJAAAAAAAAAAAAIICkgQAAAABIZWxsby50eHRQSwECFAMKAwAAAACxbPE6AAAAAAAAAAAAAAAABwAAAAAAAAAAABCA7UEzAAAAaW1hZ2VzL1BLAQIUAwoDAAAAABVn6jo8P-uKKQAAACkAAAAQAAAAAAAAAAAAIICkgVgAAABpbWFnZXMvc21pbGUuZ2lmUEsFBgAAAAADAAMAqgAAAK8AAAAAAA"},"archive_comment.zip":{"$d":"UEsDBAoAAAAAAJx08Trj5ZWwDAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsBAh4DCgAAAAAAnHTxOuPllbAMAAAADAAAAAkAAAANAAAAAAAAAKSBAAAAAEhlbGxvLnR4dGVudHJ5IGNvbW1lbnRQSwUGAAAAAAEAAQBEAAAAMwAAAAwAZmlsZSBjb21tZW50"},"backslash.zip":{"$d":"UEsDBAoAAAAAANa0J0Lj5ZWwDAAAAAwAAAAKAAAASGVsXGxvLnR4dEhlbGxvIFdvcmxkClBLAQIeAwoAAAAAANa0J0Lj5ZWwDAAAAAwAAAAKAAAAAAAAAAAAAACkgQAAAABIZWxcbG8udHh0UEsFBgAAAAABAAEAOAAAADQAAAAAAA"},"data_descriptor.zip":{"$d":"UEsDBAoACAAAAHaDIUIAAAAADAAAAAwAAAAJABwASGVsbG8udHh0VVQJAANvAONQbwDjUHV4CwABBOgDAAAE6AMAAEhlbGxvIFdvcmxkClBLBwjj5ZWwDAAAAAwAAABQSwECHgMKAAgAAAB2gyFC4-WVsAwAAAAMAAAACQAYAAAAAAAAAAAApIEAAAAASGVsbG8udHh0VVQFAANvAONQdXgLAAEE6AMAAAToAwAAUEsFBgAAAAABAAEATwAAAF8AAAAAAA"},"deflate-stream.zip":{"$d":"UEsDBBQACAAIANRxmDwAAAAAAAAAAF4AAAAJAAAASGVsbG8udHh0VcixDYAwDATAnil-AgZgDhaA-EMsJTaKLWV9aq68s2ngQnd3e1C1EwcWYaQgHUEiGyFaKyetEDdzkfbrRPHxTkaoGwazucS-fVBLBwiTTw7hSQAAAF4AAABQSwECHgMUAAgACADUcZg8k08O4UkAAABeAAAACQAAAAAAAAABAAAApIEAAAAASGVsbG8udHh0UEsFBgAAAAABAAEANwAAAIAAAAAAAA"},"deflate.zip":{"$d":"UEsDBBQAAAAIANRxmDyTTw7hSQAAAF4AAAAJAAAASGVsbG8udHh0VcixDYAwDATAnil-AgZgDhaA-EMsJTaKLWV9aq68s2ngQnd3e1C1EwcWYaQgHUEiGyFaKyetEDdzkfbrRPHxTkaoGwazucS-fVBLAQIeAxQAAAAIANRxmDyTTw7hSQAAAF4AAAAJAAAAAAAAAAEAAACkgQAAAABIZWxsby50eHRQSwUGAAAAAAEAAQA3AAAAcAAAAAAA"},"empty.zip":{"$d":"UEsFBgAAAAAAAAAAAAAAAAAAAAAAAA"},"encrypted.zip":{"$d":"UEsDBAoACQAAAJx08Trj5ZWwGAAAAAwAAAAJAAAASGVsbG8udHh0ZKbbzp8l8bsw1CrLuSqaDlU5NyMODPVMUEsHCOPllbAYAAAADAAAAFBLAQIeAwoACQAAAJx08Trj5ZWwGAAAAAwAAAAJAAAAAAAAAAAAAACkgQAAAABIZWxsby50eHRQSwUGAAAAAAEAAQA3AAAATwAAAAAA"},"extra_attributes.zip":{"$d":"UEsDBAoAAAAAAJx08Trj5ZWwDAAAAAwAAAAJABwASGVsbG8udHh0VVQJAANocGBKaHBgSnV4CwABBOgDAAAE6AMAAEhlbGxvIFdvcmxkClBLAQIeAwoAAAAAAJx08Trj5ZWwDAAAAAwAAAAJABgAAAAAAAAAAACkgQAAAABIZWxsby50eHRVVAUAA2hwYEp1eAsAAQToAwAABOgDAABQSwUGAAAAAAEAAQBPAAAATwAAAAAA"},"folder.zip":{"$d":"UEsDBAoDAAAAAANt8zoAAAAAAAAAAAAAAAAHAAAAZm9sZGVyL1BLAQIUAwoDAAAAAANt8zoAAAAAAAAAAAAAAAAHAAAAAAAAAAAAEIDtQQAAAABmb2xkZXIvUEsFBgAAAAABAAEANQAAACUAAAAAAA"},"image.zip":{"$d":"UEsDBAoDAAAAABVv6jo8P-uKKQAAACkAAAAJAAAAc21pbGUuZ2lmR0lGODdhBQAFAIACAAAAAP_eACwAAAAABQAFAAACCIwPkWerClIBADtQSwECFAMKAwAAAAAVb-o6PD_riikAAAApAAAACQAAAAAAAAAAACCApIEAAAAAc21pbGUuZ2lmUEsFBgAAAAABAAEANwAAAFAAAAAAAA"},"local_encoding_in_name.zip":{"$d":"UEsDBBQAAAAAAOabjUYAAAAAAAAAAAAAAAAMAAAAja6ioO8gr6CvqqAvUEsDBAoAAAAAAOabjUYAAAAAAAAAAAAAAAAoAAAAja6ioO8gr6CvqqAvja6i66kg4qWq4eKuouupIKSuquOspa3iLnR4dFBLAQI_ABQAAAAAAOabjUYAAAAAAAAAAAAAAAAMACQAAAAAAAAAEAAAAAAAAACNrqKg7yCvoK-qoC8KACAAAAAAAAEAGAAdZjdAB3bQAR1mN0AHdtABgTXEPQd20AFQSwECPwAKAAAAAADmm41GAAAAAAAAAAAAAAAAKAAkAAAAAAAAACAAAAAqAAAAja6ioO8gr6CvqqAvja6i66kg4qWq4eKuouupIKSuquOspa3iLnR4dAoAIAAAAAAAAQAYAB1mN0AHdtABHWY3QAd20AEdZjdAB3bQAVBLBQYAAAAAAgACANgAAABwAAAAAAA"},"nested.zip":{"$d":"UEsDBAoAAAAAAJx08Trj5ZWwDAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsDBAoAAAAAAO8GfD9R7yPjgAAAAIAAAAASAAAAemlwX3dpdGhpbl96aXAuemlwUEsDBAoAAAAAAJx08Trj5ZWwDAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsBAh4DCgAAAAAAnHTxOuPllbAMAAAADAAAAAkAAAAAAAAAAAAAAKSBAAAAAEhlbGxvLnR4dFBLBQYAAAAAAQABADcAAAAzAAAAAABQSwECHgMKAAAAAACcdPE64-WVsAwAAAAMAAAACQAAAAAAAAAAAAAApIEAAAAASGVsbG8udHh0UEsBAh4DCgAAAAAA7wZ8P1HvI-OAAAAAgAAAABIAAAAAAAAAAAAAAKSBMwAAAHppcF93aXRoaW5femlwLnppcFBLBQYAAAAAAgACAHcAAADjAAAAAAA"},"nested_data_descriptor.zip":{"$d":"UEsDBAoACAAAAJCeIUIAAAAAxAAAAMQAAAATABwAZGF0YV9kZXNjcmlwdG9yLnppcFVUCQADbzDjUG8w41B1eAsAAQToAwAABOgDAABQSwMECgAIAAAAdoMhQgAAAAAMAAAADAAAAAkAHABIZWxsby50eHRVVAkAA28A41BvAONQdXgLAAEE6AMAAAToAwAASGVsbG8gV29ybGQKUEsHCOPllbAMAAAADAAAAFBLAQIeAwoACAAAAHaDIULj5ZWwDAAAAAwAAAAJABgAAAAAAAAAAACkgQAAAABIZWxsby50eHRVVAUAA28A41B1eAsAAQToAwAABOgDAABQSwUGAAAAAAEAAQBPAAAAXwAAAAAAUEsHCLOA90TEAAAAxAAAAFBLAQIeAwoACAAAAJCeIUKzgPdExAAAAMQAAAATABgAAAAAAAAAAACkgQAAAABkYXRhX2Rlc2NyaXB0b3IuemlwVVQFAANvMONQdXgLAAEE6AMAAAToAwAAUEsFBgAAAAABAAEAWQAAACEBAAAAAA"},"nested_zip64.zip":{"$d":"UEsDBC0AAAAAALSdIUJBMtzz__________8JADAAemlwNjQuemlwVVQJAAPULuNQ1C7jUHV4CwABBOgDAAAE6AMAAAEAEAAgAQAAAAAAACABAAAAAAAAUEsDBC0AAAAAAHaDIULj5ZWw__________8JADAASGVsbG8udHh0VVQJAANvAONQbwDjUHV4CwABBOgDAAAE6AMAAAEAEAAMAAAAAAAAAAwAAAAAAAAASGVsbG8gV29ybGQKUEsBAh4DLQAAAAAAdoMhQuPllbAMAAAA_____wkAJAAAAAAAAAAAAKSBAAAAAEhlbGxvLnR4dFVUBQADbwDjUHV4CwABBOgDAAAE6AMAAAEACAAMAAAAAAAAAFBLBgYsAAAAAAAAAB4DLQAAAAAAAAAAAAEAAAAAAAAAAQAAAAAAAABbAAAAAAAAAGMAAAAAAAAAUEsGBwAAAAC-AAAAAAAAAAEAAABQSwUGAAAAAAEAAQBbAAAA_____wAAUEsBAh4DLQAAAAAAtJ0hQkEy3PMgAQAA_____wkAJAAAAAAAAAAAAKSBAAAAAHppcDY0LnppcFVUBQAD1C7jUHV4CwABBOgDAAAE6AMAAAEACAAgAQAAAAAAAFBLBgYsAAAAAAAAAB4DLQAAAAAAAAAAAAEAAAAAAAAAAQAAAAAAAABbAAAAAAAAAHcBAAAAAAAAUEsGBwAAAADSAQAAAAAAAAEAAABQSwUGAAAAAAEAAQBbAAAA_____wAA"},"pile_of_poo.zip":{"$d":"UEsDBAoAAAgAANa0fEWNgP7jIgAAACIAAAAmAAAAScOxdMOrcm7DonRpw7Ruw6BsaXrDpnRpw7hu4piD8J-SqS50eHRJw7F0w6tybsOidGnDtG7DoGxpesOmdGnDuG7imIPwn5KpUEsBAh4DCgAACAAA1rR8RY2A_uMiAAAAIgAAACYAAAAiAAAAAAAAAKSBAAAAAEnDsXTDq3Juw6J0acO0bsOgbGl6w6Z0acO4buKYg_CfkqkudHh0ScOxdMOrcm7DonRpw7Ruw6BsaXrDpnRpw7hu4piD8J-SqVBLBQYAAAAAAQABAHYAAABmAAAAIgBJw7F0w6tybsOidGnDtG7DoGxpesOmdGnDuG7imIPwn5Kp"},"slashes_and_izarc.zip":{"$d":"UEsDBBQAAAAAALtwLkLvj5iLDQAAAA0AAAAOAAAAdGVzdC9IZWxsby50eHRIZWxsbyB3b3JsZA0KUEsBAhQAFAAAAAAAu3AuQu-PmIsNAAAADQAAAA4AAAAAAAAAAQAgAAAAAAAAAHRlc3RcSGVsbG8udHh0UEsFBgAAAAABAAEAPAAAADkAAAAAAA"},"smile.gif":{"$d":"R0lGODdhBQAFAIACAAAAAP_eACwAAAAABQAFAAACCIwPkWerClIBADs"},"store-stream.zip":{"$d":"UEsDBAoACAAAANRxmDwAAAAAAAAAAAAAAAAJAAAASGVsbG8udHh0VGhpcyBhIGxvb29uZyBmaWxlIDogd2UgbmVlZCB0byBzZWUgdGhlIGRpZmZlcmVuY2UgYmV0d2VlbiB0aGUgZGlmZmVyZW50IGNvbXByZXNzaW9uIG1ldGhvZHMuClBLBwiTTw7hXgAAAF4AAABQSwECHgMKAAgAAADUcZg8k08O4V4AAABeAAAACQAAAAAAAAAAAAAApIEAAAAASGVsbG8udHh0UEsFBgAAAAABAAEANwAAAJUAAAAAAA"},"store.zip":{"$d":"UEsDBAoAAAAAANRxmDyTTw7hXgAAAF4AAAAJAAAASGVsbG8udHh0VGhpcyBhIGxvb29uZyBmaWxlIDogd2UgbmVlZCB0byBzZWUgdGhlIGRpZmZlcmVuY2UgYmV0d2VlbiB0aGUgZGlmZmVyZW50IGNvbXByZXNzaW9uIG1ldGhvZHMuClBLAQIeAwoAAAAAANRxmDyTTw7hXgAAAF4AAAAJAAAAAAAAAAAAAACkgQAAAABIZWxsby50eHRQSwUGAAAAAAEAAQA3AAAAhQAAAAAA"},"subfolder.zip":{"$d":"UEsDBAoDAAAAAANt8zoAAAAAAAAAAAAAAAAHAAAAZm9sZGVyL1BLAwQKAwAAAACbsyY7AAAAAAAAAAAAAAAAEQAAAGZvbGRlci9zdWJmb2xkZXIvUEsBAhQDCgMAAAAAA23zOgAAAAAAAAAAAAAAAAcAAAAAAAAAAAAQgO1BAAAAAGZvbGRlci9QSwECFAMKAwAAAACbsyY7AAAAAAAAAAAAAAAAEQAAAAAAAAAAABCA7UElAAAAZm9sZGVyL3N1YmZvbGRlci9QSwUGAAAAAAIAAgB0AAAAVAAAAAAA"},"text.zip":{"$d":"UEsDBAoDAAAAAJx08Trj5ZWwDAAAAAwAAAAJAAAASGVsbG8udHh0SGVsbG8gV29ybGQKUEsBAhQDCgMAAAAAnHTxOuPllbAMAAAADAAAAAkAAAAAAAAAAAAggKSBAAAAAEhlbGxvLnR4dFBLBQYAAAAAAQABADcAAAAzAAAAAAA"},"utf8.zip":{"$d":"UEsDBAoAAAAAABh9RUBq-phBBgAAAAYAAAAKAAAAYW1vdW50LnR4dOKCrDE1ClBLAQIeAwoAAAAAABh9RUBq-phBBgAAAAYAAAAKAAAAAAAAAAAAAACkgQAAAABhbW91bnQudHh0UEsFBgAAAAABAAEAOAAAAC4AAAAAAA"},"utf8_in_name.zip":{"$d":"UEsDBAoAAAgAABh9RUBq-phBBgAAAAYAAAAJAAAA4oKsMTUudHh04oKsMTUKUEsBAh4DCgAACAAAGH1FQGr6mEEGAAAABgAAAAkAAAAAAAAAAAAAAKSBAAAAAOKCrDE1LnR4dFBLBQYAAAAAAQABADcAAAAtAAAAAAA"},"winrar_utf8_in_name.zip":{"$d":"UEsDBAoAAAAAABh9RUBq-phBBgAAAAYAAAAHABIAXzE1LnR4dHVwDgABhPuNTeKCrDE1LnR4dOKCrDE1ClBLAQIfAAoAAAAAABh9RUBq-phBBgAAAAYAAAAHABIAAAAAAAAAIAAAAAAAAABfMTUudHh0dXAOAAGE-41N4oKsMTUudHh0UEsFBgAAAAABAAEARwAAAD0AAAAAAA"},"zip64.zip":{"$d":"UEsDBC0AAAAAAHaDIULj5ZWw__________8JADAASGVsbG8udHh0VVQJAANvAONQbwDjUHV4CwABBOgDAAAE6AMAAAEAEAAMAAAAAAAAAAwAAAAAAAAASGVsbG8gV29ybGQKUEsBAh4DLQAAAAAAdoMhQuPllbAMAAAA_____wkAJAAAAAAAAAAAAKSBAAAAAEhlbGxvLnR4dFVUBQADbwDjUHV4CwABBOgDAAAE6AMAAAEACAAMAAAAAAAAAFBLBgYsAAAAAAAAAB4DLQAAAAAAAAAAAAEAAAAAAAAAAQAAAAAAAABbAAAAAAAAAGMAAAAAAAAAUEsGBwAAAAC-AAAAAAAAAAEAAABQSwUGAAAAAAEAAQBbAAAA_____wAA"},"zip64_appended_bytes.zip":{"$d":"UEsDBC0AAAAAAHaDIULj5ZWw__________8JADAASGVsbG8udHh0VVQJAANvAONQbwDjUHV4CwABBOgDAAAE6AMAAAEAEAAMAAAAAAAAAAwAAAAAAAAASGVsbG8gV29ybGQKUEsBAh4DLQAAAAAAdoMhQuPllbAMAAAA_____wkAJAAAAAAAAAAAAKSBAAAAAEhlbGxvLnR4dFVUBQADbwDjUHV4CwABBOgDAAAE6AMAAAEACAAMAAAAAAAAAFBLBgYsAAAAAAAAAB4DLQAAAAAAAAAAAAEAAAAAAAAAAQAAAAAAAABbAAAAAAAAAGMAAAAAAAAAUEsGBwAAAAC-AAAAAAAAAAEAAABQSwUGAAAAAAEAAQBbAAAA_____wAASGVsbG8gV29ybGQK"},"zip64_missing_bytes.zip":{"$d":"bGxvLnR4dFVUCQADbwDjUG8A41B1eAsAAQToAwAABOgDAAABABAADAAAAAAAAAAMAAAAAAAAAEhlbGxvIFdvcmxkClBLAQIeAy0AAAAAAHaDIULj5ZWwDAAAAP____8JACQAAAAAAAAAAACkgQAAAABIZWxsby50eHRVVAUAA28A41B1eAsAAQToAwAABOgDAAABAAgADAAAAAAAAABQSwYGLAAAAAAAAAAeAy0AAAAAAAAAAAABAAAAAAAAAAEAAAAAAAAAWwAAAAAAAABjAAAAAAAAAFBLBgcAAAAAvgAAAAAAAAABAAAAUEsFBgAAAAABAAEAWwAAAP____8AAA"}}',
-  reviver,
-);
+export default JSON.parse(JSON.stringify(fixtures), reviver);
